src/aiBackend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-aiplatform==1.38.1
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
